# Added Confirmation service for modals

+++ DatingApp.Client/src/app/Components/messages/messages.component.ts
@@ -7,13 +7,14 @@ import { FormsModule } from '@angular/forms';
 import { RouterLink } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { PaginationModule } from 'ngx-bootstrap/pagination';
+import { ConfirmService } from '../../services/confirm.service';
 
 @Component({
-  selector: 'app-messages',
-  standalone: true,
-  imports: [FormsModule, ButtonsModule, RouterLink, CommonModule, PaginationModule],
-  templateUrl: './messages.component.html',
-  styleUrl: './messages.component.css'
+    selector: 'app-messages',
+    standalone: true,
+    imports: [FormsModule, ButtonsModule, RouterLink, CommonModule, PaginationModule],
+    templateUrl: './messages.component.html',
+    styleUrl: './messages.component.css'
 })
 
 export class MessagesComponent implements OnInit {
@@ -25,13 +26,13 @@ export class MessagesComponent implements OnInit {
     pageSize: number = 5;
     loadingFlag = true;
 
-    constructor(private messageService: MessageService){ }
+    constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
     ngOnInit(): void {
         this.loadMessages();
     }
 
-    loadMessages(){
+    loadMessages() {
         this.loadingFlag = true;
         this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(resp => {
             this.messages = resp.result!;
@@ -40,22 +41,25 @@ export class MessagesComponent implements OnInit {
         })
     }
 
-    pageChanged(event: any){
-        if(this.pageNumber !== event.page){
+    pageChanged(event: any) {
+        if (this.pageNumber !== event.page) {
             this.pageNumber = event.page;
             this.loadMessages();
         }
     }
 
-    getLink(message: MessageModel)
-    {
+    getLink(message: MessageModel) {
         var output = this.container === "Outbox" ? "/members/" + message.recipientUsername : "/members/" + message.senderUsername;
         return output;
     }
 
-    deleteMessage(id: number){
-        this.messageService.deleteMessage(id).subscribe(() =>{
-            this.messages.splice(this.messages.findIndex(x => x.id === id));
-        });
+    deleteMessage(id: number) {
+        this.confirmService.confirm('Confirm Delete message', 'This cannot be undone...').subscribe(result => {
+            if (result) {
+                this.messageService.deleteMessage(id).subscribe(() => {
+                    this.messages.splice(this.messages.findIndex(x => x.id === id));
+                });
+            }
+        })
     }
 }
